+++ app/src/App.jsx
@@ -1,39 +1,26 @@
+import Loading from 'layouts/Loading';
+import Tool from 'layouts/Tool';
 import { useState } from 'react';
 import { useEffectOnce } from 'react-use';
 import { useSetRecoilState } from 'recoil';
 import { paletteState, themeState } from 'states';
-import { Tool, Loading } from 'layouts';
-import { getColors } from 'utils';
+import { getCSSColors } from 'utils';
 
 export default () => {
   const setTheme = useSetRecoilState(themeState);
   const setPalette = useSetRecoilState(paletteState);
-  const [isReady, setReady] = useState(false);
+  const [isDataLoaded, setDataLoaded] = useState(false);
 
   useEffectOnce(() => {
-    setTheme(getColors());
-    const observer = new MutationObserver(() => setTheme(getColors()));
+    setTheme(getCSSColors());
+    const observer = new MutationObserver(() => setTheme(getCSSColors()));
     observer.observe(document.documentElement, { attributes: true, attributeFilter: ['style'] });
   });
 
   useEffectOnce(() => {
-    window.addEventListener('message', e => {
-      const { type, payload } = e.data;
-      switch (type) {
-        case 'SET_INIT_VALUES':
-          setPalette(payload.palette);
-          setReady(true);
-          break;
-
-        case 'SET_PALETTE':
-          setPalette(payload);
-          break;
-
-        default:
-          break;
-      }
-    });
+    setPalette(window.$palette);
+    setDataLoaded(true);
   });
 
-  return <div className='select-none'>{isReady ? <Tool /> : <Loading />}</div>;
+  return <div className='select-none'>{isDataLoaded ? <Tool /> : <Loading />}</div>;
 };