+++ app/src/components/Elements/index.jsx
@@ -1,6 +1,6 @@
-import Element from './Element';
-import { isExpandState } from 'states';
 import { useRecoilState } from 'recoil';
+import { isExpandState } from 'states';
+import Element from './Element';
 
 export default ({ children }) => {
   const [isExpand, setExpand] = useRecoilState(isExpandState(children[0].title));