+++ app/src/layouts/index.jsx
@@ -1,4 +1,0 @@
-import Tool from './Tool';
-import Loading from './Loading';
-
-export { Tool, Loading };