+++ app/src/utils.js
@@ -1,6 +1,14 @@
+import { colord } from 'colord';
+import palettes from 'data/palettes';
+import merge from 'lodash/fp/merge';
+
 const vscode = window.acquireVsCodeApi();
 
-export const getColors = () => {
+export function postMessage(msg) {
+  return vscode.postMessage(msg);
+}
+
+export function getCSSColors() {
   const props = document.documentElement.style.cssText.split(';');
   const colors = props.reduce((style, property) => {
     let [name, value] = property.split(':');
@@ -8,15 +16,40 @@ export const getColors = () => {
     return { ...style, [name]: value };
   }, {});
   return colors;
-};
+}
 
-// export const setGlobals = () => {
-//   window.$shrink = {};
-//   const colors = Object.entries({ base: palettes.base, ...palettes[window.$palette] });
-//   window.$colors = colors.map(([name, indents]) => [name, Object.entries(indents)]);
-// };
+export function getColorValues(value) {
+  const Colord = colord(value);
+  return {
+    rgb: Colord.toRgb(),
+    hsl: Colord.toHsl(),
+    hex: { hex: Colord.toHex(), a: Colord.alpha() },
+  };
+}
 
-export const postMessage = msg => vscode.postMessage(msg);
+export function getColorString(value, format) {
+  const Colord = colord(value);
+  switch (format) {
+    case 'hex':
+      return Colord.toHex();
+
+    case 'hsl':
+      return Colord.toHslString();
+
+    default:
+      return Colord.toRgbString();
+  }
+}
+
+export function getHexFromValues({ a, hex, ...rest }) {
+  return colord(hex || rest)
+    .alpha(a)
+    .toHex();
+}
+
+export function getColorsFromPalette({ type, colors }) {
+  return Object.entries(merge(palettes[type], colors)).map(([name, indents]) => [name, Object.entries(indents)]);
+}
 
 export const getFallbackScope = name => {
   const [scope, _name] = name.split('_');
@@ -24,5 +57,3 @@ export const getFallbackScope = name => {
   if (name[0] !== '$' || scopes.length < 2) return '';
   return `${scopes.slice(0, scopes.length - 1).join('.')}_${_name}`;
 };
-
-export const toColorsArray = colors => Object.entries(colors).map(([name, indents]) => [name, Object.entries(indents)]);