+++ extension/handleMessage.js
@@ -1,19 +1,20 @@
+const config = require('./utils/config');
 const handlePreviews = require('./handlePreviews');
 const handleThemeChange = require('./handleThemeChange');
 
 module.exports = ({ type, payload }) => {
   console.log('💬', payload);
-
   switch (type) {
-    case 'SET_THEME':
-      handleThemeChange(payload);
-      // theme = createTheme({ theme, ...payload });
-      // fs.write(paths.theme, theme, true);
-      break;
-    case 'OPEN_PREVIEW':
+    case 'PREVIEW':
       handlePreviews(payload);
       break;
-    default:
+
+    case 'SET_THEME':
+      handleThemeChange(payload);
+      break;
+
+    case 'SET_PALETTE_COLORS':
+      config.set('vslook.palette.colors', payload);
       break;
   }
 };