+++ extension/index.js
@@ -1,30 +1,35 @@
 const vscode = require('vscode');
 const config = require('./utils/config');
-const palettes = require('./palettes');
 const createWebview = require('./createWebview');
 const handleMessage = require('./handleMessage');
 
+const getTheme = () => {
+  // const tokenColors = config
+  //   .get('editor.tokenColorCustomizations.textMateRules', [])
+  //   .map(({ scope, settings }) => Object.entries(settings).map(([name, value]) => [`$${scope}_${name}`, value]))
+  //   .flat();
+
+  const colorCustomizations = config.get('workbench.colorCustomizations');
+
+  return colorCustomizations;
+
+  // return { ...colorCustomizations, ...Object.fromEntries(tokenColors) };
+};
+
+const getPalette = () => ({
+  type: config.get('vslook.palette.type').toLowerCase(),
+  colors: config.get('vslook.palette.colors'),
+});
+
 function activate(context) {
-  let webview;
   const paths = { root: context.extensionPath };
-  const palette = config.get('vslook.palette').toLowerCase();
 
   const disposableOnEdit = vscode.commands.registerCommand(`vslook.edit`, () => {
-    webview = createWebview({ root: paths.root }, handleMessage);
-    webview.postMessage({
-      type: 'SET_INIT_VALUES',
-      payload: { palette: { base: palettes.base, ...palettes[palette] } },
-    });
+    const data = { palette: getPalette(), theme: getTheme() };
+    createWebview({ root: paths.root, data }, handleMessage);
   });
 
-  const disposableOnConfigChange = vscode.workspace.onDidChangeConfiguration(() => {
-    if (!!webview) {
-      const palette = config.get('vslook.palette').toLowerCase();
-      webview.postMessage({ type: 'SET_PALETTE', payload: { base: palettes.base, ...palettes[palette] } });
-    }
-  });
-
-  context.subscriptions.push(disposableOnConfigChange, disposableOnEdit);
+  context.subscriptions.push(disposableOnEdit);
 }
 
 function deactivate() {}